feat(playwright-ct-angular): ✨ align with other fmwks: props/on
